robust CSD estimation, restructured and added gradient for optimization, which significantly speeds up the speed (1s using the analytical gradient vs. 7s using gradiant)
